+++ AdNetworkSupport/Chartboost/ChartboostRewardedVideoCustomEvent.h
@@ -12,7 +12,7 @@
 #endif
 
 /*
- * Certified with version 6.4.0 of the Chartboost SDK.
+ * Certified with version 6.5.2 of the Chartboost SDK.
  */
 
 @interface ChartboostRewardedVideoCustomEvent : MPRewardedVideoCustomEvent