+++ AdNetworkSupport/Millennial/MPMillennialBannerCustomEvent.m
@@ -1,3 +1,9 @@
+//
+//  MPMillennialBannerCustomEvent.m
+//
+//  Copyright (c) 2015 Millennial Media, Inc. All rights reserved.
+//
+
 #import "MPMillennialBannerCustomEvent.h"
 #import "MPLogging.h"
 #import "MPAdConfiguration.h"
@@ -7,110 +13,93 @@
 #define MM_SIZE_300x250 CGSizeMake(300, 250)
 #define MM_SIZE_728x90  CGSizeMake(728, 90)
 
-@interface MPInstanceProvider (MillennialBanners)
-
-- (MMInlineAd *)buildMMInlineAdWithSize:(CGSize)size placementId:(NSString *)placementId;
-
-@end
+static NSString *const kMoPubMMAdapterAdUnit = @"adUnitID";
+static NSString *const kMoPubMMAdapterDCN = @"dcn";
 
 @implementation MPInstanceProvider (MillennialBanners)
 
-- (MMInlineAd *)buildMMInlineAdWithSize:(CGSize)size placementId:(NSString *)placementId
-{
+- (MMInlineAd *)buildMMInlineAdWithSize:(CGSize)size placementId:(NSString *)placementId {
     return [[MMInlineAd alloc] initWithPlacementId:placementId size:size];
 }
 
 @end
 
-////////////////////////////////////////////////////////////////////////////////////////////////////
-
 @interface MPMillennialBannerCustomEvent ()
 
 @property (nonatomic, assign) BOOL didTrackImpression;
 @property (nonatomic, assign) BOOL didTrackClick;
-@property (nonatomic, assign) BOOL didShowModal;
 
-@property (nonatomic, strong) MMInlineAd *mmInlineAds;
-
-- (CGRect)frameFromCustomEventInfo:(NSDictionary *)info;
-- (CGSize)sizeFromCustomEventInfo:(NSDictionary *)info;
+@property (nonatomic, strong) MMInlineAd *mmInlineAd;
 
 @end
 
 @implementation MPMillennialBannerCustomEvent
 
-@synthesize didTrackImpression = _didTrackImpression;
-@synthesize didTrackClick = _didTrackClick;
-@synthesize mmInlineAds = _mmInlineAds;
-
-
-- (BOOL)enableAutomaticImpressionAndClickTracking
-{
+- (BOOL)enableAutomaticImpressionAndClickTracking {
     return NO;
 }
 
-- (id)init
-{
+- (id)init {
     self = [super init];
     if (self) {
-        if ( ![[MMSDK sharedInstance] isInitialized] ) {
-            [[MMSDK sharedInstance] initializeWithSettings:[[MMAppSettings alloc] init] withUserSettings:[[MMUserSettings alloc] init]];
+        if ([[UIDevice currentDevice] systemVersion].floatValue >= 7.0) {
+            MMSDK *mmSDK = [MMSDK sharedInstance];
+            if ([mmSDK isInitialized] == NO) {
+                MMAppSettings *appSettings = [[MMAppSettings alloc] init];
+                [mmSDK initializeWithSettings:appSettings withUserSettings:nil];
+            }
         }
     }
     return self;
 }
 
-- (void)invalidate
-{
-    self.mmInlineAds = nil;
+- (void)dealloc {
+    self.mmInlineAd = nil;
     self.delegate = nil;
 }
 
-- (void)dealloc
-{
-    [self invalidate];
-}
+- (void)requestAdWithSize:(CGSize)size customEventInfo:(NSDictionary *)info {
 
-- (void)requestAdWithSize:(CGSize)size customEventInfo:(NSDictionary *)info
-{
-    MPLogInfo(@"Requesting Millennial banner");
-    NSString *placementId = [info objectForKey:@"adUnitID"];
+    MMSDK *mmSDK = [MMSDK sharedInstance];
 
-    // If Custom Event Class Data contains DCN / Position, let's use that instead.
-    // { "dcn": "...", "adUnitID": "..."  }
-    [[MMSDK sharedInstance] appSettings].mediator = @"MPMillennialBannerCustomEvent";
-    if ( [info objectForKey:@"dcn"] ) {
-        [[[MMSDK sharedInstance] appSettings] setSiteId:[info objectForKey:@"dcn"]];
-    } else {
-        [[[MMSDK sharedInstance] appSettings] setSiteId:nil];
+    if ([mmSDK isInitialized] == NO) {
+        MPLogError(@"Millennial adapter not properly intialized yet.");
+        [self.delegate bannerCustomEvent:self didFailToLoadAdWithError:nil];
+        return;
     }
 
-    if ( !placementId ) {
+    MPLogInfo(@"Requesting Millennial banner.");
+
+    NSString *placementId = info[kMoPubMMAdapterAdUnit];
+    if (!placementId) {
         MPLogError(@"Millennial received invalid placement ID. Request failed.");
         [self.delegate bannerCustomEvent:self didFailToLoadAdWithError:nil];
         return;
     }
 
-    [[MMSDK sharedInstance] setSendLocationIfAvailable:[[MoPub sharedInstance] locationUpdatesEnabled]];
+    [mmSDK appSettings].mediator = @"MPMillennialBannerCustomEvent";
+    if (info[kMoPubMMAdapterDCN]) {
+        [mmSDK appSettings].siteId = info[kMoPubMMAdapterDCN];
+    } else {
+        [mmSDK appSettings].siteId = nil;
+    }
 
-    self.mmInlineAds = [[MPInstanceProvider sharedProvider] buildMMInlineAdWithSize:size placementId:placementId];
-    self.mmInlineAds.delegate = self;
-    self.mmInlineAds.refreshInterval = -1;
+    self.mmInlineAd = [[MPInstanceProvider sharedProvider] buildMMInlineAdWithSize:size placementId:placementId];
+    self.mmInlineAd.delegate = self;
+    self.mmInlineAd.refreshInterval = -1;
 
-    [self.mmInlineAds.view setFrame:[self frameFromCustomEventInfo:info]];
-    [self.mmInlineAds request:nil];
+    [self.mmInlineAd.view setFrame:[self frameFromCustomEventInfo:info]];
+    [self.mmInlineAd request:nil];
 
 }
 
-- (CGSize)sizeFromCustomEventInfo:(NSDictionary *)info
-{
-    CGFloat width = [[info objectForKey:@"adWidth"] floatValue];
-    CGFloat height = [[info objectForKey:@"adHeight"] floatValue];
+- (CGSize)sizeFromCustomEventInfo:(NSDictionary *)info {
+    CGFloat width = [info[@"adWidth"] floatValue];
+    CGFloat height = [info[@"adHeight"] floatValue];
     return CGSizeMake(width, height);
 }
 
-- (CGRect)frameFromCustomEventInfo:(NSDictionary *)info
-{
+- (CGRect)frameFromCustomEventInfo:(NSDictionary *)info {
     CGSize size = [self sizeFromCustomEventInfo:info];
     if (!CGSizeEqualToSize(size, MM_SIZE_300x250) && !CGSizeEqualToSize(size, MM_SIZE_728x90)) {
         size.width = MM_SIZE_320x50.width;
@@ -119,38 +108,34 @@
     return CGRectMake(0, 0, size.width, size.height);
 }
 
-
 #pragma mark - MMInlineAdDelegate methods
 
 - (UIViewController *)viewControllerForPresentingModalView {
     return [self.delegate viewControllerForPresentingModalView];
 }
 
-
 - (void)inlineAdContentTapped:(MMInlineAd *)ad {
     if (!self.didTrackClick) {
-        MPLogInfo(@"Millennial banner was clicked.");
+        MPLogInfo(@"Millennial banner %@ was clicked.", ad);
         [self.delegate trackClick];
         self.didTrackClick = YES;
     } else {
-        MPLogInfo(@"Millennial banner ignoring duplicate click");
+        MPLogInfo(@"Millennial banner %@ ignoring duplicate click.", ad);
     }
 }
 
-
 - (void)inlineAdWillPresentModal:(MMInlineAd *)ad {
-    MPLogInfo(@"Millennial banner will present modal");
-    self.didShowModal = YES;
+    MPLogInfo(@"Millennial banner %@ will present modal.", ad);
     [self.delegate bannerCustomEventWillBeginAction:self];
 }
 
 - (void)inlineAdDidCloseModal:(MMInlineAd *)ad {
-    MPLogInfo(@"Millennial banner did dismiss modal");
+    MPLogInfo(@"Millennial banner %@ did dismiss modal.", ad);
     [self.delegate bannerCustomEventDidFinishAction:self];
 }
 
 - (void)inlineAdRequestDidSucceed:(MMInlineAd *)ad {
-    MPLogInfo(@"Millennial banner did load");
+    MPLogInfo(@"Millennial banner %@ did load.", ad);
     [self.delegate bannerCustomEvent:self didLoadAd:ad.view];
     [self didDisplayAd];
     if (!self.didTrackImpression) {
@@ -160,8 +145,8 @@
 }
 
 - (void)inlineAd:(MMInlineAd *)ad requestDidFailWithError:(NSError *)error {
-    MPLogError(@"Millennial interstitial ad failed with error (%d) %@", error.code, error.description);
-    [self.delegate bannerCustomEvent:self didFailToLoadAdWithError:nil];
+    MPLogError(@"Millennial banner %@ failed with error (%d) %@", ad, error.code, error.description);
+    [self.delegate bannerCustomEvent:self didFailToLoadAdWithError:error];
 }