+++ AdNetworkSupport/Millennial/MPMillennialInterstitialCustomEvent.h
@@ -1,18 +1,23 @@
+//
+//  MPMillennialInterstitialCustomEvent.h
+//
+//  Copyright (c) 2015 Millennial Media, Inc. All rights reserved.
+//
 
 #if __has_include(<MoPub/MoPub.h>)
     #import <MoPub/MoPub.h>
 #else
-    #import "MPInterstitialCustomEvent.h"
     #import "MoPub.h"
 #endif
 
 #import <MMAdSDK/MMAdSDK.h>
-#import <MMAdSDK/MMInterstitialAd.h>
 
 /*
- * For MMSDK version 6.0
+ * For MMSDK version 6.3.1
  */
 
 @interface MPMillennialInterstitialCustomEvent : MPInterstitialCustomEvent <MMInterstitialDelegate>
+
+@property (nonatomic, readonly) MMInterstitialAd *interstitial;
 
 @end