+++ AdNetworkSupport/TapJoy/TapjoyRewardedVideoCustomEvent.m
@@ -1,8 +1,7 @@
 
 #import "TapjoyRewardedVideoCustomEvent.h"
-
-#import <Tapjoy/TJPlacement.h>
 #import <Tapjoy/Tapjoy.h>
+#import <Tapjoy/TJPlacement.h>
 #import "MPRewardedVideoError.h"
 #import "MPLogging.h"
 #import "MPRewardedVideoReward.h"
@@ -11,28 +10,82 @@
 
 @interface TapjoyRewardedVideoCustomEvent () <TJPlacementDelegate, TJCVideoAdDelegate>
 @property (nonatomic, strong) TJPlacement *placement;
+@property (nonatomic, assign) BOOL isConnecting;
+@property (nonatomic, copy) NSString *placementName;
 @end
 
 @implementation TapjoyRewardedVideoCustomEvent
 
-- (void)requestRewardedVideoWithCustomEventInfo:(NSDictionary *)info
-{
-    MPLogInfo(@"Requesting Tapjoy rewarded video");
+- (void)setupListeners {
+    [[NSNotificationCenter defaultCenter] addObserver:self
+                                             selector:@selector(tjcConnectSuccess:)
+                                                 name:TJC_CONNECT_SUCCESS
+                                               object:nil];
+
+    [[NSNotificationCenter defaultCenter] addObserver:self
+                                             selector:@selector(tjcConnectFail:)
+                                                 name:TJC_CONNECT_FAILED
+                                               object:nil];
+}
+
+- (void)initializeWithCustomNetworkInfo:(NSDictionary *)info {
     //Instantiate Mediation Settings
     TapjoyGlobalMediationSettings *medSettings = [[MoPub sharedInstance] globalMediationSettingsForClass:[TapjoyGlobalMediationSettings class]];
 
-    if (![Tapjoy isConnected]) {
-        [Tapjoy connect:medSettings.sdkKey
-                options:medSettings.connectFlags];
+    // Grab sdkKey and connect flags defined in MoPub dashboard
+    NSString *sdkKey = info[@"sdkKey"];
+    BOOL enableDebug = [info[@"debugEnabled"] boolValue];
+
+    if (medSettings.sdkKey) {
+        MPLogInfo(@"Connecting to Tapjoy via MoPub mediation settings");
+        [self setupListeners];
+        [Tapjoy connect:medSettings.sdkKey options:medSettings.connectFlags];
+
+        self.isConnecting = YES;
+
     }
+    else if (sdkKey) {
+        MPLogInfo(@"Connecting to Tapjoy via MoPub dashboard settings");
+        NSMutableDictionary *connectOptions = [[NSMutableDictionary alloc] init];
+        [connectOptions setObject:@(enableDebug) forKey:TJC_OPTION_ENABLE_LOGGING];
+        [self setupListeners];
+
+        [Tapjoy connect:sdkKey options:connectOptions];
+
+        self.isConnecting = YES;
+    }
+    else {
+        MPLogInfo(@"Tapjoy rewarded video is initialized with empty 'sdkKey'. You must call Tapjoy connect before requesting content.");
+        [self.delegate rewardedVideoDidFailToLoadAdForCustomEvent:self error:nil];
+    }
+}
+
+- (void)requestRewardedVideoWithCustomEventInfo:(NSDictionary *)info {
     // Grab placement name defined in MoPub dashboard as custom event data
-    NSString *name = info[@"name"];
+    self.placementName = info[@"name"];
 
-    if(name) {
-        _placement = [TJPlacement placementWithName:name mediationAgent:@"mopub" mediationId:nil delegate:self];
-        _placement.adapterVersion = @"3.0";
+    // Adapter is making connect call on behalf of publisher, wait for success before requesting content.
+    if (self.isConnecting) {
+        return;
+    }
 
-        [_placement requestContent];
+    // Attempt to establish a connection to Tapjoy
+    if (![Tapjoy isConnected]) {
+        [self initializeWithCustomNetworkInfo:info];
+    }
+    // Live connection to Tapjoy already exists; request the ad
+    else {
+        MPLogInfo(@"Requesting Tapjoy rewarded video");
+        [self requestPlacementContent];
+    }
+}
+
+- (void)requestPlacementContent {
+    if (self.placementName) {
+        self.placement = [TJPlacement placementWithName:self.placementName mediationAgent:@"mopub" mediationId:nil delegate:self];
+        self.placement.adapterVersion = MP_SDK_VERSION;
+
+        [self.placement requestContent];
     }
     else {
         MPLogInfo(@"Invalid Tapjoy placement name specified");
@@ -41,32 +94,27 @@
     }
 }
 
-- (void)presentRewardedVideoFromViewController:(UIViewController *)viewController
-{
+- (void)presentRewardedVideoFromViewController:(UIViewController *)viewController {
     if ([self hasAdAvailable]) {
         MPLogInfo(@"Tapjoy rewarded video will be shown");
-        [_placement showContentWithViewController:nil];
+        [self.placement showContentWithViewController:nil];
     }
     else {
         MPLogInfo(@"Failed to show Tapjoy rewarded video");
         NSError *error = [NSError errorWithDomain:MoPubRewardedVideoAdsSDKDomain code:MPRewardedVideoAdErrorNoAdsAvailable userInfo:nil];
         [self.delegate rewardedVideoDidFailToPlayForCustomEvent:self error:error];
     }
-
 }
 
-- (BOOL)hasAdAvailable
-{
-    return _placement.isContentAvailable;
+- (BOOL)hasAdAvailable {
+    return self.placement.isContentAvailable;
 }
 
-- (void)handleCustomEventInvalidated
-{
-    _placement.delegate = nil;
+- (void)handleCustomEventInvalidated {
+    self.placement.delegate = nil;
 }
 
-- (void)handleAdPlayedForCustomEventNetwork
-{
+- (void)handleAdPlayedForCustomEventNetwork {
     // If we no longer have an ad available, report back up to the application that this ad expired.
     // We receive this message only when this ad has reported an ad has loaded and another ad unit
     // has played a video for the same ad network.
@@ -75,12 +123,13 @@
     }
 }
 
-- (void)dealloc
-{
+- (void)dealloc {
+    [[NSNotificationCenter defaultCenter] removeObserver:self];
     _placement.delegate = nil;
 }
 
 #pragma mark - TJPlacementDelegate methods
+
 - (void)requestDidSucceed:(TJPlacement *)placement {
     if (!placement.isContentAvailable) {
         MPLogInfo(@"No Tapjoy rewarded videos available");
@@ -93,6 +142,7 @@
     MPLogInfo(@"Tapjoy rewarded video content is ready");
     [self.delegate rewardedVideoDidLoadAdForCustomEvent:self];
 }
+
 - (void)requestDidFail:(TJPlacement *)placement error:(NSError *)error {
     MPLogInfo(@"Tapjoy rewarded video request failed");
     [self.delegate rewardedVideoDidFailToLoadAdForCustomEvent:self error:error];
@@ -119,6 +169,18 @@
     [self.delegate rewardedVideoShouldRewardUserForCustomEvent:self reward:[[MPRewardedVideoReward alloc] initWithCurrencyAmount:@(kMPRewardedVideoRewardCurrencyAmountUnspecified)]];
 }
 
+- (void)tjcConnectSuccess:(NSNotification*)notifyObj {
+    MPLogInfo(@"Tapjoy connect Succeeded");
+    self.isConnecting = NO;
+    [self requestPlacementContent];
 
+    [[NSNotificationCenter defaultCenter] removeObserver:self];
+}
+
+- (void)tjcConnectFail:(NSNotification*)notifyObj {
+    MPLogInfo(@"Tapjoy connect Failed");
+    self.isConnecting = NO;
+    [[NSNotificationCenter defaultCenter] removeObserver:self];
+}
 
 @end